RPCN v0.6

+++ rpcs3/Emu/Cell/Modules/sceNp.h
@@ -1393,14 +1393,14 @@ struct SceNpMatchingReqRange
 
 struct SceNpScoreVariableSizeGameInfo
 {
-	be_t<u64> infoSize;
+	be_t<u32> infoSize;
 	u8 data[SCE_NP_SCORE_VARIABLE_SIZE_GAMEINFO_MAXSIZE];
 	u8 pad[3];
 };
 
 struct SceNpScoreRecordOptParam
 {
-	be_t<u64> size;
+	be_t<u32> size;
 	vm::bptr<SceNpScoreVariableSizeGameInfo> vsInfo;
 	vm::bptr<CellRtcTick> reserved;
 };
